flounder: Disable forced encryption on data by default
Change to encryptable on f2fs and ext4

It turns out there are some Nexus 9s out there that have /data
formatted to ext4, so the change has to be applied on both ext4 and
f2fs.

There are some users that have disabled encryption on their device.
By forcing encryption, the first boot will re-encrypt his/her device.
Otherwise, it gives the user the choice in settings

Change-Id: Ie1c3c99a8ab2867b5c837f1a420e3e1a145a895b